Working implementation of Hardware Interface for MqGateway #13 This is still work-in-progress, but working - tested on MqGateway without expander (4 MCP23017s)

+++ src/main/kotlin/com/mqgateway/core/hardware/mqgateway/MqGatewayConnector.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 data class MqGatewayConnector(
   val portNumber: Int,
   val wireColor: WireColor,
-  val debounceMs: Int?
+  val debounceMs: Long?
 ) : HardwareConnector
 
 enum class WireColor(val number: Int) {
